now the payment handler will take the number passed in by the body turns it into a payment intent the returns it in a response body

+++ backend/internal/models/stripe.go
@@ -2,6 +2,6 @@ package model
 
 type (
 	Stripe struct {
-		Amount float64 `json:"amount"`
+		Amount int64 `json:"amount"`
 	}
 )
